Basic Salary
David needs to write a program to calculate the net payable salary using Inheritance, Constructors and this pointer. The program will take basic salary as an 
input from the user and calculate the total salary along with the following components: 
HRA, which is 50% of the basic salary 
Special allowance, which is 75% of the basic salary 
PF deduction. which is 12% of the basic salary 
Ater calculating the preceding salary components, the net payable salary should be displayed. Help David to 
perform the preceding requirement.

 

Example

Input :   30000

output:

Basic salary = 30000
HRA = 15000
Special allowance = 22500
PF = 3600
Net Payable salary = 63900